move poincare_details test to Library (large data test doesn't work)
